Easier detection of SparkPlan nodes causing GPU conversion failures
